Updated Sri's like update
Fixed a bug where if the user is logged in, likes a name and then refreshes the page - the "filled" heart goes back to default.
Now it checks the logged in user's likedbabyname list from the database and checks if the name is liked.
So the state stays persistent even after the page is refreshed.